WalletInfo: add findSignerByAddress and refactor requests to use it
Reduces code duplication in Checkout / SignTransaction / SignTransactionLedger by moving
logic for determing signer for a sender address to WalletInfo. The requests do now NOT
handle unsupported HTLC contracts anymore (an unhandled exception is thrown) which is not
an issue though as currently the contract finding does not return any HTLC accounts and
most probably will not until we actually have HTLC support.