Prevent parallel calls to Network to result in the second failing
The second fails because the NetworkClient.Instance is already set after the first call, but internally the NetworkClient is still awaiting the RpcClient's init() call. But because the instance is already set, the Network.vue component does not wait for the second request, throwing "this._eventClient is undefined" or "Cannot read property call of undefined" errors.