Network: fix initialization of Network client
Fixes a bug introduced in https://github.com/nimiq/hub/pull/312
that results in calls to instances of the Network component to never
execute for instances that are not the first created. This bug
occured as the promise waiting for the client initialization was
not shared across Network component instances.