Send credentials with the webhook (#345)
To support various e-commerce authentication systems, we allow to send credentials (cookies and auth headers) with the webhook request. However, because including the credentials requires the server to answer with a specific header (as per CORS rules), this might fail if the server omits this header. We then simply try again with the credentials omitted.