package.json: upgrade typescript to support built-in Omit
Also updating @types/node to avoid error "Duplicate identifier 'IteratorResult'"
between newest typescript and previous @types/node. Also added @types/node to
package.json as it was previously only installed as a sub dependency of other
dependencies but is actually a direct dependency as we're using node's 'require',
'global' and 'process' at some places in the code. Made it a dev dependency
as no node types are required in the output code or types.