RequestTypes,*PaymentOptions: sanitize currency and type
Check that the currency and type of the unparsed input payment options
object matches the ParsedPaymentOptions it should get parsed to. While
for the initial request data this is already ensured by the RequestParser
this is relevant for updated data that comes in via the webhook.

To access the currency and type of the child classes already in the
constructor of the base class, these have been changed from properties
to getters. Additionally, some common properties have been moved from
the base class to the interface.