Update and customize browser warnings
Update to browser warning 1.1.0 which now provides a share button,
fixes issues with old browsers, updated the warning for Edge
browsers, adds types and is minified.
The share button is customized such that it shares the caller url
instead of the hub itself which wouldn't make sense as the hub
request would not be available anymore when opening in another
tab or browser. Cashlinks however are shared normally.