Checkout: make carousel full height on mobile
Previously, only the second step, after a currency was selected, was
full height. Now also the first step with the currency infos over the
carousel is full height which follows the designs and transitions
smoother to the second step.
Additionally, now do not apply the carousel full height anymore on
medium size screens that do not use the mobile layout.