RequestParser: don't strip unselected checkout currencies from the request
Stripping unselected currencies from the request had undesired side effects
on reload: When the user selects a currency and then reloads the page, the
not-selected currencies got stripped from the request. On second reload,
the modified request is loaded from the history state by RpcApi and rejected
if another currency than NIM is the only one that remained in the request as
checkout requests need to contain a NIM option.