Upgrade @nimiq/rpc to fix a bug that prevents any requests in Firefox
After #397 for some reason a Promise polyfill is active in Firefox. This caused a problem
in the rpc lib which could not detect a native Promise as a Promise it needs to await
anymore. As a consequence it tried to resolve each request immediately with the promise
as result which lets redirect requests return immediately and popup requests fail as
promises can not be sent over postMessage.