LedgerSwapProxy: make the proxy a multisig
Add the Ledger as secondary signer to avoid that funds become inaccesible
if the secret salt is lost. This way, funds locked on the proxy can be
redeemed using the Ledger. Funds locked on the HTLC are temporarily
inaccesible, as the Ledger Nimiq app currently does not support signing
HTLC transactions, but will be accesible once HTLC support is added to
the Ledger app.

This change also maintains backwards compatibility to legacy, un-salted
proxies. Compatibility to salted, non-multisig proxies implemented in the
last commits does not have to be maintained, as they were never deployed.