Cashlink: simplify state detection
Also now in theory more resilient to very long cashlink tx histories
that can not be retrieved entirely from nodes such that knownFundingTx
is not included anymore. In this case the detection is now able to at
least reach CHARGED or CLAIMING (which is also displayed in the ui
similarly to CLAIMED) states.