HubApi: BrowserStack local iOS testing compatibility
On iOS, BrowserStack routes localhost connections for testing of local code
through an actual domain (bs-local.com) because iOS blocks localhost connections
to content that's not sitting on the phone.
The HubApi now uses bs-local:8080 instead of localhost:8080 by default, if the
main app is running on bs-local.com, such that a locally served dev server can
be tested with BrowserStack.