vue.config.js: avoid webpack caching external BitcoinJS as undefined
Because we load BitcoinJS only asynchronously on demand, we have to apply
a hack to avoid Webpack caching the dependency while it's still undefined.
Instead, we ensure the export is always re-evaluated to the current global
variable BitcoinJS by overwriting the exports of the cached module as getter.

The reason this issue only surfaced now is likely that previously we only
triggered webpack's import of BitcoinJS after we already loaded the bundle
which makes the global variable available, but now @ethersproject also imports
'buffer' which is mapped to the version bundled with BitcoinJS.