Update @nimiq/utils and @nimiq/vue-components for new FiatApi; and make FiatApi provider configurable
The updated FiatApi supports CryptoCompare as an alternative API provider and
supports bridged exchange rates for currencies not directly supported by the
providers.

CoinGecko is becoming increasingly strict with unauthenticated access to their
API without API key. For example, wallet.nimiq.com has already been blocked from
unauthenticated access, thus the Wallet has been switched to CryptoCompare as
fiat api provider, while hub.nimiq.com and keyguard.nimiq.com are not currently
blocked. I.e. in the Hub, we don't have to switch yet. If we switch the Hub's
FiatApi provider in PaymentInfoLine, we should also switch it in the Keyguard's
PaymentInfoLine, which is a bit more work though. For this reason, we keep the
Hub and Keyguard on CoinGecko for now, but remove the CoinGecko proxy / custom
API URL, as this one has been shut down.