Update @nimiq/ledger-api to 3.0.0
Update to latest @nimiq/ledger-api which includes support for, among others,
core-rs-albatross and Albatross transactions, also staking transactions,
compatibility with Nimiq Ledger app 2.0, dependency updates etc.

For now, the API is configured in this commit to work with legacy core-js and
legacy transactions.

In order to make the API work with the pretty outdated Webpack and typescript
versions used in the Hub, some patch work was necessary via patch-package and
babel transpilation via transpileDependencies.