@nimiq/ledger-api: avoid Webpack trying to bundle dynamically loaded Albatross core wasm
Patch @nimiq/ledger-api to avoid the Hub's Webpack trying to process and bundle the dynamically loaded
Albatross core wasm, which is intended to be loaded from the copy created in vue.config.js.

Also change Hub's dev port tested for in ledger-api from 8081 to correct 8080. This will be fixed
in the next @nimiq/ledger-api release without requiring a patch.