@nimiq/ledger-api: quick-fix patch to adapt path to Hub's @nimiq/core files
Patch @nimiq/ledger-api to adapt the path to where the Hub's @nimiq/core files are located.
This quick-fix currently hardcodes @nimiq/core's version number, which will be addressed in
a follow up change. The version number should be either adapted automatically (in the patch
or the Ledger API), or ideally, @nimiq/core should become an external dependency in the
Ledger API, and then be injected via the Hub's vue.config.js