Polygon,ethers: support for Polygon network and gas-abstracted transactions
This includes connecting to the Polygon network and getting infos like balances,
and block height, sending gas-abstracted transactions via OpenGSN and Uniswap,
getting fees for gas-abstracted transactions, fetching transaction histories for
gas-abstracted transactions etc.
Signing of gas-abstracted transactions is not supported yet, and these libraries
are not used anywhere yet. Next steps will include adding support for signing
and using the libraries in Cashlinks.

Primarily, this adapts and refactors the libraries ethers.ts, OpenGSN.ts and
Uniswap.ts from the Wallet for use in the Hub. Eventually, the goal will be
to create them in a separate repository as a shared library between the Wallet
and Hub.