KeyReplacer,vue.config.js: support providing API keys in a secure fashion
Testnet and mainnet API keys are fetched from the server, in the same fashion as the Wallet does it. This
avoids them being exposed in the repository at least. Testnet and mainnet keys typically can and should
employ tighter privileges compared to dev keys, restricting them to use only from the deployed domains.

Dev keys live in an env file as alternative mechanism to fetching the key from the server. This has the
advantage especially for development keys that they don't have to sit in the /public folder as fetchable
file, which would end up in the builds and deployments even for testnet and mainnet build targets. This is
important as development keys can be especially powerful, compared to other keys, as they typically allow
access from localhost or any domain.
The env file must be listed in .gitignore to avoid it being exposed in the repository and must be included
only in dev builds, to avoid the key still becoming exposed as part of the app build for other build targets.
Thus, it's recommendable to keep the keys in a .env.development.local file, which is already configured that
way by the Vue CLI, see https://cli.vuejs.org/guide/mode-and-env.html#environment-variables.
Note for dev builds, the key is then included in the app build, and could be exposed that way, but typically,
dev builds are not publicly deployed, such that this shouldn't be a problem.